move base classes to template
