+++ src/classes/Caches.ts
@@ -1,5 +1,5 @@
 import { Invite } from '@uoa-discords/shared-utils';
-import CachedStorage from './CachedStored';
+import CachedStorage from './template/CachedStored';
 
 export default abstract class Caches {
     public static readonly inviteCache = new CachedStorage<Invite>();
