+++ src/classes/template/CachedStored.ts
