+++ src/classes/template/Logger.ts
